% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bank_RmvLMM}
\alias{bank_RmvLMM}
\title{Aggregate Group-wise RmvLMM Results using Divide-and-combine strategy}
\usage{
bank_RmvLMM(rds_files, indep_snp_file, out_file = NULL)
}
\arguments{
\item{rds_files}{A character vector of file paths to the .rds files generated by \code{run_RmvLMM}.}

\item{indep_snp_file}{File path to the list of approximately independent SNPs. The file must contain a column named 'SNP'.}

\item{out_file}{Optional. File path to save the final calibrated results (.rds).}
}
\value{
A list containing the calibrated P-value data.frame and averaged variance components (Vb, Ve).
}
\description{
Reads multiple .rds files generated by \code{run_RmvLMM}, performs Fisher's omnibus
combination, and calibrates P-values using a Generalized Gamma distribution based on independent SNPs.
}
