% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fit.generalgamma.mle}
\alias{fit.generalgamma.mle}
\title{Maximum likelihood estimation of the generalized gamma distribution}
\usage{
fit.generalgamma.mle(data, inits = NULL)
}
\description{
Maximum likelihood estimation of the generalized gamma distribution
}
