% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fit_multivariate_lmm_optimized}
\alias{fit_multivariate_lmm_optimized}
\title{Fit a multivariate linear mixed model}
\usage{
fit_multivariate_lmm_optimized(
  precomputed_data,
  max_iter = 200,
  tol = 1e-04,
  verbose = 0
)
}
\description{
Fit a multivariate linear mixed model
}
