% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{reml_neg_log_lik_optimized}
\alias{reml_neg_log_lik_optimized}
\title{Univariate REML negative log-likelihood}
\usage{
reml_neg_log_lik_optimized(xi, y, M, Ak_diag)
}
\description{
Univariate REML negative log-likelihood
}
